+++ src/utils/hooks/api/loans/use-get-account-loans-overview.tsx
@@ -4,20 +4,13 @@ import Big from 'big.js';
 import { useCallback } from 'react';
 
 import { convertMonetaryAmountToValueInUSD } from '@/common/utils/utils';
-import { GOVERNANCE_TOKEN } from '@/config/relay-chains';
 import { BorrowAction, LendAction } from '@/types/loans';
 import { getTokenPrice } from '@/utils/helpers/prices';
 
 import useAccountId from '../../use-account-id';
 import { useGetPrices } from '../use-get-prices';
-import { getPositionsTotalUSDField } from './get-usd-values';
 import { useGetAccountPositions } from './use-get-account-positions';
-
-const getTotalEarnedRewards = (lendPositions: LendPosition[] = [], borrowPositions: BorrowPosition[] = []) =>
-  [...lendPositions, ...borrowPositions].reduce(
-    (total, position) => (position.earnedReward ? total.add(position.earnedReward) : total),
-    newMonetaryAmount(0, GOVERNANCE_TOKEN)
-  );
+import { getPositionsSumOfFieldsInUSD, getTotalEarnedRewards } from './utils';
 
 interface AccountLoansOverviewData {
   lendPositions: LendPosition[] | undefined;
@@ -67,15 +60,15 @@ const useGetAccountLoansOverview = (): AccountLoansOverview => {
   let collateralRatio: number | undefined = undefined;
 
   if (lendPositions !== undefined && prices !== undefined) {
-    lentAssetsUSDValue = getPositionsTotalUSDField('amount', lendPositions, prices);
-    totalEarnedInterestUSDValue = getPositionsTotalUSDField<LendPosition>('earnedInterest', lendPositions, prices);
+    lentAssetsUSDValue = getPositionsSumOfFieldsInUSD('amount', lendPositions, prices);
+    totalEarnedInterestUSDValue = getPositionsSumOfFieldsInUSD<LendPosition>('earnedInterest', lendPositions, prices);
 
     const collateralLendPositions = lendPositions.filter(({ isCollateral }) => isCollateral);
-    collateralAssetsUSDValue = getPositionsTotalUSDField('amount', collateralLendPositions, prices);
+    collateralAssetsUSDValue = getPositionsSumOfFieldsInUSD('amount', collateralLendPositions, prices);
   }
 
   if (borrowPositions !== undefined && prices !== undefined) {
-    borrowedAssetsUSDValue = getPositionsTotalUSDField('amount', borrowPositions, prices);
+    borrowedAssetsUSDValue = getPositionsSumOfFieldsInUSD('amount', borrowPositions, prices);
   }
 
   if (borrowedAssetsUSDValue !== undefined && collateralAssetsUSDValue !== undefined) {
